+++ test/Ocelot.AcceptanceTests/LoadBalancer/LoadBalancerTests.cs
@@ -107,7 +107,7 @@ public sealed class LoadBalancerTests : ConcurrentSteps, IDisposable
             _services = services;
         }
 
-        public async Task<Response<ServiceHostAndPort>> Lease(HttpContext httpContext)
+        public async Task<Response<ServiceHostAndPort>> LeaseAsync(HttpContext httpContext)
         {
             var services = await _services();
             lock (_lock)