+++ test/Ocelot.UnitTests/LoadBalancer/DelegateInvokingLoadBalancerCreatorTests.cs
@@ -107,7 +107,7 @@ namespace Ocelot.UnitTests.LoadBalancer
             public DownstreamRoute DownstreamRoute { get; }
             public IServiceDiscoveryProvider ServiceDiscoveryProvider { get; }
 
-            public Task<Response<ServiceHostAndPort>> Lease(HttpContext httpContext)
+            public Task<Response<ServiceHostAndPort>> LeaseAsync(HttpContext httpContext)
             {
                 throw new NotImplementedException();
             }