+++ ReleaseNotes.md
@@ -1,1 +1,8 @@
 Technical release, version {0}
+
+### Breaking changes
+
+- The `ILoadBalancer` interface: The `Lease` method was renamed to `LeaseAsync`.
+  Interface FQN: `Ocelot.LoadBalancer.LoadBalancers.ILoadBalancer`
+  Method FQN:  `Ocelot.LoadBalancer.LoadBalancers.ILoadBalancer.LeaseAsync`
+- TO BE Written